Broadcom STB: add buildroot config script
Add br_config.pl to simplify configuring buildroot for Broadcom STB.

Signed-off-by: Markus Mayer <mmayer@broadcom.com>

commit a00630dcc6d598f36afa0390e3505d17ef6968b6
Author: Markus Mayer <mmayer@broadcom.com>
Date:   Thu Feb 22 12:29:27 2018 -0800

br_config.pl: fix warning about $arch being undefined

Ensure that we access $arch only when it is defined, otherwise we'll
get a warning if br_config.pl is called without arguments. To do so,
we move early accesses to $arch a bit further down the script, where we
know for certain that it is defined.

refs #SWLINUX-4756

Signed-off-by: Markus Mayer <mmayer@broadcom.com>