system/Config.in: add configuration options for 32-bit library support
In order to support 32-bit shared libraries in a 64-bit root file
system, we need several configuration options that let us tell
buildroot which libraries to put where.

Specifically, we introduce the following:

BR2_ROOTFS_LIB_DIR
------------------

Tells buildroot where to put the main system libraries. On 32-bit
setups, this will mean 32-bit libraries. On 64-bit setups, this
setting will apply to 64-bit libraries. Typical values for
BR2_ROOTFS_LIB_DIR are "lib", "lib32" or "lib64".

The default value for 32-bit setups is "lib". For 64-bit setups it is
"lib64".

BR2_ROOTFS_LIB32_DIR
--------------------

This tells buildroot where to put 32-bit libraries on a 64-bit system.
Typical values for BR2_ROOTFS_LIB32_DIR are "lib" or "lib32". The
default value is "lib".

On 32-bit setups, this variable is not available. On 64-bit systems
with 32-bit support disabled, this variable will not be used.

BR2_ROOTFS_RUNTIME32
--------------------

Tells buildroot whether it should install 32-bit shared libraries into
a 64-bit root file system. If set to false, the root file system will
not be able to support 32-bit binaries.

This variable is not available on 32-bit setups.

BR2_ROOTFS_RUNTIME32_PATH
-------------------------

Points to the sys-root of a 32-bit toolchain for the current 64-bit
architecture (i.e. an ARM sys-root on AArch64 or MIPS32 on MIPS64).

This path is used to copy all 32-bit shared libraries from the 32-bit
sys-root into the buildroot staging area. From there, they will be
included into the root file system. Care is taken that 32-bit libraries
will not interfere with 64-bit libraries already in staging.

We need the ability to specify a separate sys-root for 32-bit, because
the 64-bit sys-root may not include 32-bit libraries.

This variable is not available on 32-bit setups.

Signed-off-by: Markus Mayer <mmayer@broadcom.com>