board: brcmstb: add post-image script
We introduce a post-image script which will remove rootfs.cpio as well
as some build stage files related to the Linux build, resulting in the
kernel being linked twice upon "make", once without initrd, and once
with initrd. These will be the same build stages that are also
executed on a clean build.

Signed-off-by: Markus Mayer <mmayer@broadcom.com>