board: brcmstb: post-build: copy skel/ from uclinux-rootfs
We copy the contents from the old uclinux-rootfs into our new BR
rootfs.

If we ever need to be more selective, we'll need to extract the old
skeleton into a scratch area and pick what we need from there. For now,
we simply extract the entire skeleton directly into our rootfs.

Signed-off-by: Markus Mayer <mmayer@broadcom.com>