package: add portmap_stb
Add portmap package for backward compatibility with the legacy
uclinux-rootfs.

Note: the config option is called BR2_PACKAGE_PORTMAP_STB and the
package is called portmap_stb, since a legacy option called
BR2_PACKAGE_PORTMAP already exists in Buildroot. (It references the old
portmap package that used to be part of buildroot and that has since
been removed.)

Despite the name, this is the "standard" portmap software without any
modifications.

Signed-off-by: Markus Mayer <mmayer@broadcom.com>