br_config.pl: add command line switch to disable download cache
Provide a means to tell br_config.pl that we don't want to use a shared
download cache, even if it is available.

Amongst other things, forcing downloads can be helpful for nightly
builds to detect issues that may otherwise remain hidden for a period
of time. It'll also help us find software packages that need to be added
to our local buildroot mirror.

refs #SWLINUX-4764

Signed-off-by: Markus Mayer <mmayer@broadcom.com>