br_config.pl: detect toolchain kernel header version
Buildroot insists on being configured with the kernel header version
matching what's included in the toolchain.

Our toolchain may include different kernel header versions for
different platforms. Therefore, we want to detect at configuration time
which kernel header version is bundled with the the toolchain for the
selected platform. This version number is then incorporated into the BR
configuration, so Buildroot won't complain.

refs #SWLINUX-4796

Signed-off-by: Markus Mayer <mmayer@broadcom.com>