board: brcmstb: Make GPLv3 check work again
We have a check for GPLv3 packages in /etc/init.d/rcS which just checks
for gdb/gdbserver. There are actually more GPLv3 packages included by
default within buildroot, and now that we compile such a list using
"make legal-info, we can just check that this list exists and is
non-zero.

refs #SWLINUX-4731

Signed-off-by: Florian Fainelli <florian.fainelli@broadcom.com>