br_config.pl: fix kernel header version detection
The regex that generates the toolchain base directory by stripping a
trailing "/bin", if present, was too restrictive and wouldn't match
"/bin/". This would lead to the wrong base directory being used to find
the sysroot, which would return an empty string.

Ultimately, br_config.pl would evaluate the host system's Linux
version, which would lead to kernel version mismatches with Buildroot
and would cause the build to fail.

In addition to the improved regex, we now also detect an empty sysroot
and return "undef", so we don't accidentally evaluate the host's Linux
version in the future.

Lastly, we fix the regex mentioned above in two locations, the second
location being 32-bit support for 64-bit platforms. We need to the
toolchain base directory there, as well.

refs #SWLINUX-4796

Signed-off-by: Markus Mayer <mmayer@broadcom.com>