br_config.pl: force BR to re-sync the toolchain after running br_config.pl
Every time br_config.pl executes, it looks for a toolchain to use. If
the $PATH changes or a toolchain is specified via "-t", it could be
different from the last time br_config.pl was run. However, by default,
Buildroot only sets up the toolchain for a clean build and keeps
re-using the configured toolchain. As a result, br_config.pl and
Buildroot might have different ideas as to which toolchain will be used
for the subsequent build.

To ensure that Buildroot is actually using the toolchain br_config.pl
thinks it will use, we have br_config.pl clean out the currently
configured toolchain and have it trigger Buildroot to set up its
toolchain symlinks again when the next build is started.

refs #SWLINUX-4827

Signed-off-by: Markus Mayer <mmayer@broadcom.com>