linux: turn on INITRD only for initrd-rebuild
If we enable INITRD by default and merely supply an empty initrd for
the non-initrd kernel, the resulting kernel will not boot properly. In
order to resolve this issue, we need to ensure to actually turn off the
INITRD configuration option for the initial kernel build and enable it
only for the initrd-rebuild.

refs #SWLINUX-4861

Signed-off-by: Markus Mayer <mmayer@broadcom.com>