br_config.pl: Exclude all *.o files
When using the same Linux source tree between uclinux-rootfs and
buildroot, it is possible to create a situation where we produce *.o
files for a given architecture, e.g: ARM, and then attempt an ARM64
build, this can lead to build errors for tools like tmon which would not
rebuild the existing *.o files.

Extend the exclusion list to cover all *.o files, which would also take
care of testing for vmlinux.o in check_linux().

refs #SWLINUX-4855

Signed-off-by: Florian Fainelli <florian.fainelli@broadcom.com>