board: brcmstb: be more diligent detecting the kernel directory
Our post-build scripts need to access the kernel directory that was
used for the build.

We need to be more strict in our logic of finding the right directory
(so we don't use "linux-tools" or "linux-firmware" as presumed kernel
directory), while at taking into account that the directory might be
called "linux-custom" or "linux-stb-...".

refs #SWLINUX-4903

Signed-off-by: Markus Mayer <mmayer@broadcom.com>