board: brcmstb: post-build: abort build if stbtools can't be found
We abort the build if we can't find the stbtools tar-ball. Without it,
we can't extract our custom "skel" files. As a result, all kinds of
things will break in unexpected and confusing ways at run time.

It's better to throw an error at build time, so everybody knows the
build result wouldn't have worked as intended.

refs #SWLINUX-4992

Signed-off-by: Markus Mayer <mmayer@broadcom.com>