br_config.pl: handle kernel configuration "defconfig" properly
While not all architectures use "defconfig" as the default
configuration name, many do. Also, Buildroot expects a custom kernel
configuration to be specified without trailing "defconfig". That means
specifying just "defconfig" as the kernel defconfig name leads to an
empty defconfig name in the Buildroot configuration -- which will not
work.

Therefore, we treat the kernel defconfig name "defconfig" as a special
case. If it is specified, we assume the user meant to use the
platform's default configuration and proceed accordingly.

refs #SWLINUX-5065

Signed-off-by: Markus Mayer <mmayer@broadcom.com>