br_config.pl: attempt to fix directory permission in shared DL cache
Directories in the shared download cache need to be world-writable, so that any
developer can download new software tar-balls (which happens automatically
during the build process).

Since we have been running into permission problems whenever the above criteria
are not met, br_config.pl attempts to fix directory permissions in the shared
download cache if it detects that the shared download cache exists. This will
only work for those directories with wrong permissions that are owned by the
caller of the script.

While this may not eliminate all permission problems right away, it can be
hoped that, with time, every developer who owns directories in the shared
download cache will run br_config.pl eventually, thus resolving the issue
with time.

refs #SWLINUX-5085

Signed-off-by: Markus Mayer <mmayer@broadcom.com>