board: brcmstb: post-build: improve logic that searches for stbtools
The introduction of config variable BR2_DL_DIR_OPTS breaks the
logic searching for the stbtools tar-ball in the shared download cache.

We used to grep for "BR2_DL_DIR" in the .config. Unfortunately,
BR2_DL_DIR_OPTS matches this search in addition to the match we want.
This results in stbtools*.tar.gz not being found and the build
aborting.

To prevent this issue, we now grep for "BR2_DL_DIR=", ensuring we get
only one search result.

refs #SWLINUX-5094

Signed-off-by: Markus Mayer <mmayer@broadcom.com>