board: brcmstb: post-build: make ldd available on the target
Having ldd available on the target can be very helpful for diagnosing
issues with shared libraries. Since ldd is installed as part of the
host tools, but not in the target, we simply copy the host's ldd script
in to the target rootfs (it's a bash script, so this works).

refs #SWLINUX-5099

Signed-off-by: Markus Mayer <mmayer@broadcom.com>