build_rootfs_images.sh: check for availability of file system tools
We check if the file system utilities we need to generate various file
system images are actually present on the build machine before calling
them.

The utilities we need are: mksquashfs, mkfs.ubifs and mkfs.jffs2

We now detect if one of these tools is missing, issue a warning and
skip the file system format in question. In the past, we attempted to
invoke the tool anyway. This would then fail with a less obvious
error, making it harder to determine the root cause why certain file
system formats were not present after the build.

refs #SWLINUX-5111

Signed-off-by: Markus Mayer <mmayer@broadcom.com>