misc: Makefile.brcmstb: explicitly check if br_config.pl failed
We need to check the exit code of br_config.pl within the for-loop that
builds the architecture images. If br_config.pl terminates with an
error, we need to abort the build.

This is explicit check is necessary, because the for-loop hides error
codes of individual iterations from the make command. Subsequent
iterations of the loop will simply overwrite the exit code of previous
iterations. The exit code of the final iteration is returned to make
as exit code of the for-loop as a whole. This means that an
architecture may fail to build properly, but the build process will
still terminate successfully, as long as the last architecture builds
correctly.

refs #SWLINUX-5122

Signed-off-by: Markus Mayer <mmayer@broadcom.com>