br_config.pl: check if local Linux copy is stale
If a user lets br_config.pl pick the default kernel, the build process
will clone the Linux repo from stbgit.broadcom.com and create a local
copy of said kernel sources. This local copy will then be used for
subsequent builds.

Since the build process only checks for the existence of a local copy
of the kernel, but not its age, it'll continue using the existing copy
without ever updating it. Over time, the local copy of the Linux
kernel can become quite outdated.

To prevent this from happening, we let br_config.pl check if the local
copy of the kernel sources is older than seven days. It'll look in the
local "dl/" directory or in the shared download cache if present. If it
finds that the Linux sources are too old, it'll remove them, thus
forcing the build process to create a fresh clone from the server.

refs #SWLINUX-5180

Signed-off-by: Markus Mayer <mmayer@broadcom.com>