bin: br_config: perform basic sanity check of build environment
We want br_config.pl to perform a basic sanity check of the build
environment. If it finds troublesome settings that are known to cause
issues, it'll abort.

For now we check that:
   - "." is not in the PATH, because Buildroot's build will error out if
     it is.
   - /tools/bin is not in the PATH, because that is known to cause run
     time problems for host tools on older Linux installations

We also make a minor tweak to fix_oss_permissions(), so br_config.pl will
work with older Perl versions.

Lastly, we simplify get_kernel_header_version() a little.

refs #SWLINUX-5210

Signed-off-by: Markus Mayer <mmayer@broadcom.com>