busybox: tftp: create files with execute permissions
Busybox's TFTP client has been modified by Broadcom (in 2013) to create
new transferred files with execute permissions. The original code
change can be found at:
http://stbgit.broadcom.com/cgit/38.git/commit/?id=c9b98d6ca74ae298a3b6ad2b9c8e476bebade540

We re-create this behaviour in our buildroot rootfs, but this time as a
user configurable option rather than by simply hardcoding it. In
addition and for consistency, we're not just changing the TFTP client,
but also the TFTP server.

refs #SWLINUX-5229

Signed-off-by: Markus Mayer <mmayer@broadcom.com>