package/mtd: ensure we build with crypto disabled on the host
We disable crypto support for mtd-utils.

With crypto disabled we must also pass "-DNDEBUG" to the C compiler.
Otherwise crypto-related assertions are being triggered, which cause
mkfs.ubifs to fail.

This is only needed as of mtd-utils v2.1.0, which is being used in
Buildroot 2019.05.

Tweaked for mtd-utils v2.1.1, which is being used from Buildroot 2019.08
onwards.

refs #SWLINUX-5254

Signed-off-by: Markus Mayer <mmayer@broadcom.com>