package/dosfstools: introduce sanitation step during install
We can't install dosfstools into the target directory directly, because
it'll install *all* binaries, even the disabled ones.

To avoid this issue, we install everything into a temporary location
first and remove those binaries that have been disabled in the
Buildroot configuration before copying the rest to the target location.

refs #SWLINUX-5256

Signed-off-by: Markus Mayer <mmayer@broadcom.com>