board: brcmstb: post-build: ldd support for 32-bit binaries on 64-bit system
ldd easily supports different binary formats, but it needs to know the
shared library loader for all formats it is supposed to handle.
However, the ldd script we are copying into the root file system only
knows one shared library loader: the one for the "main" binary format
(i.e. only 32-bit or only 64-bit).

In order to inspect Aarch32 binaries on Aarch64 systems, we need to add
the shared library loader for 32-bit binaries to the list.

refs #SWLINUX-5298

Signed-off-by: Markus Mayer <mmayer@broadcom.com>