bin: makedelta.pl: script to generate Linux kernel diffs
We add a slightly modified and improved copy of makedelta.pl, which as
originally taken from 49.git/tools/makedelta.pl.

This copy of makedelta.pl takes two command line arguments:
  - the path to the Linux kernel
  - the path to the output source directory

The changes make it easier to call makedelta.pl from a makefile.

There are also some stylistic fixes and a close() call was added to
close the Makefile the script has been reading.

refs #SWLINUX-5307

Signed-off-by: Markus Mayer <mmayer@broadcom.com>