package: brcm-skel: transition skeleton installation from post-build script
We transition the responsibility to install all BRCMSTB skeleton files
from post-build.sh to package/brcm-skel.

The advantages of the new approach are:

  1. Less custom scripting. Buildroot logic takes care of most steps.
  2. It becomes possible to use BR's xxx_OVERRIDE_SRCDIR facility.

We want to be able to the xxx_OVERRIDE_SRCDIR directives for our skeleton
files to help with development, but also to let continuous integration build
custom images containing developers' changes to the Broadcom tools before
they are committed into the stbtools repo.

refs #SWLINUX-5316

Signed-off-by: Markus Mayer <mmayer@broadcom.com>