board: brcmstb: post-build: don't check for 32-bit libs in 64-bit environment
We need to ensure that 32-bit support is enabled before calling
find_64bit_libs.sh to warn about wrongly installed 32-bit libraries.
The reason is that Buildroot will create lib64 as symlinks to lib if
there is no 32-bit support in a 64-bit build.

Firstly, attempting to copy libraries from lib to lib64 is pointless if
these two directories point to the same location. Secondly, the copy
operation will actually fail because of this, causing the entire build
to fail.

So, let's not call find_64bit_libs.sh if lib64 is a symlink.

refs #SWLINUX-5347

Signed-off-by: Markus Mayer <mmayer@broadcom.com>