bin: br_config.pl: support requesting a specific toolchain version
We add a new command line option (-T) that allows the user to request a
specific toolchain version ("stbgcc-6.1-1.6" for instance). This is
helpful in ensuring reproducible builds.

Until now, br_config.pl would simply use the newest toolchain it could
find in $PATH or /opt/toolchains if there was no toolchain in the PATH.

Option -t does allow one to specify a specific toolchain, but it
requires a full path and doesn't search for a particular toolchain in
all known locations. Hence the new option, which does just that.

refs #SWLINUX-5351

Signed-off-by: Markus Mayer <mmayer@broadcom.com>