misc: Makefile.brcmstb: ensure we use the recommended toolchain
With br_config.pl's new option (-T) in place, we can now request the
toolchain specified in misc/toolchain.master or misc/toolchain. (The
former takes precedence.)

Using the recommended toolchain makes the builds more reproducible and
ensures that the same toolchain is being used that was used for
testing.

refs #SWLINUX-5351

Signed-off-by: Markus Mayer <mmayer@broadcom.com>