misc: Makefile.brcmstb: simplify Linux version retrieval
Simplify and future-proof how we determine the Linux version. The
current approach is based on, and is compatible with, the classic root
file system, but it has drawbacks. Also, there are a few unused
variables.

The new approach doesn't require continuous updating as new major Linux
versions are being released.

refs #SWLINUX-5351

Signed-off-by: Markus Mayer <mmayer@broadcom.com>