misc/Makefile.brcmstb: exclude rootfs.cpio from release
Since we are now keeping rootfs.cpio after the build, we must ensure
that we don't copy it into the release directory.

For one, rootfs.cpio does not contain any useful data that should be
part of the release. Secondly, the file is called the same for all
architectures. Attempting to copy multiple files of the same name into
a single directory would lead to a clash.

refs #SWLINUX-5377

Signed-off-by: Markus Mayer <mmayer@broadcom.com>