board: brcmstb: post-build.sh: try to detect skeleton inconsistencies
To detect inconsistencies in our skeleton files (say, if files got
overwritten by another package due to incorrect dependencies), we
compare /etc/inittab in our target root file system with the version in
the source package.

While not a fool-proof way of finding a mismatch, it stands to reason
that if another skeleton package overwrote our files, inittab would be
one of them.

refs #SWLINUX-5410

Signed-off-by: Markus Mayer <mmayer@broadcom.com>