br_config.pl: handle (lib|lib64) and (lib32|lib) library directories
The directories for 64-bit and 32-bit libraries on a 64-bit system has
changed. We introduce code that lets br_config.pl detect whether the
toolchain being used is a traditional stbgcc toolchain or a new one and
configure buildroot accordingly.

On a traditional toolchain, we have:
    /lib:   32-bit libraries
    /lib64: 64-bit libraries

Using a new toolchain, we have:
    /lib:   64-bit libraries
    /lib32: 32-bit libraries
    /lib64: symlink to /lib

refs #SWLINUX-5421

Signed-off-by: Markus Mayer <mmayer@broadcom.com>