support/download/dl-wrapper: set umask for shared download cache
When we are using a shared download cache, we have to ensure that all
directories and files we create are writable by anybody. That way it
remains possible to download updates irrespective of the user doing so.

To ensure proper file permissions (0777 for directories, 0666 for
files), we set our umask to 0 when we detect that the download
directory is our shared cache location.

refs #SWLINUX-5442

Signed-off-by: Markus Mayer <mmayer@broadcom.com>