br_config.pl: use ACLs to set the default UMASK of the download cache
We try to use ACLs to set the default UMASK in the shared download
cache to world-writable. This will work for newly created download
caches, provided the file system supports ACLs.

We also attempt to fix the ACLs for already existing download caches.
This has to be a best-effort as the user running br_config.pl may not
have the permission to set ACLs.

refs #SWLINUX-5442

Signed-off-by: Markus Mayer <mmayer@broadcom.com>