core/pkg-toolchain-external: pass BR2_ROOTFS_LIB_DIR to copy_toolchain_sysroot
We pass $(BR2_ROOTFS_LIB_DIR) to copy_toolchain_sysroot so we can
decide at copy time if the architecture's library directory and our
desired run-time library directory are different.

Making this distinction, and handling it, allows us to copy shared
libraries from the toolchain's sys-root library directory into a
different location on the target. We need this for traditional stbgcc
toolchains. Shared libraries for 64-bit architectures are stored in
/lib64 in those toolchains, but we want to copy them into /lib in our
root file system.

refs #SWLINUX-5435

Signed-off-by: Markus Mayer <mmayer@broadcom.com>