system/Config.in: introduce BR2_NEED_LD_SO_CONF
Introduce a Kconfig variable to denote if we need ld.so.conf or not.

Since we want to use /lib32 as the library directory for the 32-bit
run-time on a 64-bit system even if the toolchain has been configured
to use /lib, we need ld.so.conf to tell the shared library loader where
to find 32-bit libraries, since it won't look in /lib32 otherwise.

ld.so.conf is not necessary with a toolchain that was configured from
the outset to use /lib for 64-bit libraries and /lib32 for 32-bit
libraries.

We must also prevent the top-level Makefile from complaining when we
are required to use ld.so.conf.

refs #SWLINUX-5435

Signed-off-by: Markus Mayer <mmayer@broadcom.com>