system/Config.in: introduce pre-build script
We introduce the concept of a pre-build script that works similar to
the already existing post-build and post-image scripts.

The pre-build script(s) are executed before the build commences. This
allows a user to run some preperatory tasks prior to the build.

The arguments passed to the pre-build script are
BR2_ROOTFS_POST_SCRIPT_ARGS, i.e. the same arguments passed to
post-build and post-image!

refs #SWLINUX-5435

Signed-off-by: Markus Mayer <mmayer@broadcom.com>