board: brcmstb: add ld.so.conf and ldconfig
We add ld.so.conf and copy it into the root file system when needed.

We also add a "cross-capable" x86 version of ldconfig that can process
ARM libraries. We use it on the build host to generate ld.so.cache for
the target if required.

Lastly, we also copy the ARM64 version of ldconfig into the target file
system, so it is available should the need arise.

refs #SWLINUX-5435

Signed-off-by: Markus Mayer <mmayer@broadcom.com>