package/glibc: deal with GCC 9 for host package
For GCC 9 to be able to build GLIBC 2.30, we need to "--enable-cet".
To make things more complicated, enabling "Intel Control-flow
Enforcement Technology" requires the compiler option "-fcf-protection",
which is not supported on older compilers. So, we need to check if the
compiler knows the option and enable CET only if it does.

refs #SWLINUX-5435

Signed-off-by: Markus Mayer <mmayer@broadcom.com>