br_config.pl: support for in-line config fragments
We add command line options -R <string> and -r <string>. The former
allows specifying kernel config fragments, the latter Buildroot config
fragments. The fragment text is given directly and not via file.

  $ ./bin/br_config.pl -r '# BR2_PACKAGE_CMATOOL is not set' ... arm

Multiple config settings can be separated by semicolon:

   ... -R 'CONFIG_DEBUG_DRIVER=y;CONFIG_DEBUG_DEVRES=y' ...

This feature is meant for a small number of config options, ideally
just a single one.

The feature is especially useful for our CI setup when we need to
temporarily tweak an option or two. It greatly simplifies achieving
this, avoiding the need to handle temporary files (and even
directories) in the CI scripts.

refs #SWLINUX-5470

Signed-off-by: Markus Mayer <mmayer@broadcom.com>