br_config.pl: add support for ccache
We add support for ccache and introduce command line option "-X".

If it doesn't already exist, br_config.pl will attempt to create the
directory hierarchy for a shared compiler cache and set world writable
permissions. If this is not possible, it'll revert to configuring a
per-user compiler cache (which will not be made world writable).

The aforementioned behaviour is the default, if option "-X" is not
specified. Option "-X" can be used to specify an alternate location
for the compiler cache and can also be used to turn off using ccache
altogether ("-X -").

CAUTION! br_config.pl will *NOT* attempt to create any leading
directories if a custom cache location is specified via "-X". It is the
responsibility of the user to ensure the specified location is usable.
The user must also specify any umask values manually if so desired.

refs #SWLINUX-5513

Signed-off-by: Markus Mayer <mmayer@broadcom.com>