br_config.pl: disable CMATOOL if custom kernel doesn't support it
We detect whether a custom kernel (specified using command line option
"-L") contains the STB CMA driver required to build CMATOOL. We disable
CMATOOL if the kernel driver doesn't exist.

This does not work for kernels automatically cloned from a remote repo,
as we can't check for the existence of
    include/linux/brcmstb/cma_driver.h
at configuration time.

refs #SWLINUX-5573

Signed-off-by: Markus Mayer <mmayer@broadcom.com>