br_config.pl: better permission check for shared ccache
We have to check permissions on the sub-directories 0-9 and a-f, as
well as tmp, inside the shared ccache, if they exist. They have to be
writable to the user. Otherwise, we have to use a private ccache or
the build will fail.

Without this change, the following situation will lead to build errors.

$ ls -la /local/users/stbdev/buildroot-ccache
total 88
drwxrwxrwx 19 user1 stbdev  4096 May 11 14:15 .
drwxrwxrwx  3 user1 stbdev  4096 Mar 19 10:47 ..
drwxr-xr-x 18 user2 stbdev  4096 May  7 15:08 0
drwxr-xr-x 18 user2 stbdev  4096 May  7 15:09 1
drwxr-xr-x 18 user2 stbdev  4096 May  7 15:09 2
[...]

That's because the cache directory itself is world-writable, but the
individual sub-directories aren't. br_config.pl would not detect this
and attempt to use the shared ccache.

refs #SWLINUX-5574

Signed-off-by: Markus Mayer <mmayer@broadcom.com>