board: brcmstb: pre-build.sh: Remove kernel modules
When doing incremental kernel builds and switching
branches/tags/releases, the /lib/modules/ directory will get filed with
unique directory names that contain the kernel version as well as the
git tag. This eventually creates a large initramfs which BOLT is not
capable of loading anymore.

If a developer wishes to keep the modules around, setting
BR_KEEP_KERNEL_MODULES as an environment variable will permit that.

refs #SWLINUX-5567

Signed-off-by: Florian Fainelli <florian.fainelli@broadcom.com>