busybox: Support passing arguments to the reboot command
It is desirable to support a PMIC powercycle which is achieved by
specifying a string argument to the "reboot" command. When such an
argument is specified, we only support a forced reboot which leads
straight to the execution of the system call. The reason for that is
that the non-forced reboot would involve passing that string to the init
program which is a bit difficult without resorting to files, pipes or
sockets. It is not deemed necessary to do that at the moment.

Another approach was considered which was to use a script wrapper with a
simple C program doing the system call, but that could present some
challenges with customer deployments.

refs #SWLINUX-5575

Signed-off-by: Florian Fainelli <florian.fainelli@broadcom.com>