board: brcmstb: post-build.sh: handle musl's version of ld.so.conf
With musl-libc, the run-time file that lists shared library search
paths is called /etc/ld-musl-$ARCH.path and not /etc/ld.so.conf as is
the case with glibc. Also, musl-libc doesn't require ldconfig.

Luckily, the file format of ld.so.conf and ld-musl-$ARCH.path can be
the same, so we can just copy our existing ld.so.conf and rename it
appropriately.

https://wiki.musl-libc.org/faq.html#Q:-Where-is-%3Ccode%3Eldconfig%3C/code%3E?

refs #SWLINUX-5600

Signed-off-by: Markus Mayer <mmayer@broadcom.com>