board: brcmstb: add linux-pre-build.sh
Move the logic to remove old kernel modules from the generic pre-build
script to a kernel-specific pre-build script.

This change was necessitated by the fact that the generic pre-build
script is invoked for every target, even ones that don't actually
rebuild any code, such as "make legal-info".

Additionally, we also need to remove the ".stamp_target_installed" file
for the kernel whenever we clean up kernel modules. Cleaning up will
also remove the current kernel modules. Therefore, we need to tell the
build system that they need to be re-installed.

refs #SWLINUX-5601

Signed-off-by: Markus Mayer <mmayer@broadcom.com>