misc/Makefile.brcmstb: Fix bmips builds
The output of bin/get_archs.pl is a buildroot configuration target such as
bmips, however this was used directly without massaging for checking that an
arch/*/ directory existed which would fail since there is no arch/bmips in
Linux.

We would then also forget to check that bmips_stb_defconfig exists which is
indicative of a Broadcom STB reference kernel build.

refs #SWLINUX-5665

Fixes: adddfb06aa12 ("misc/Makefile.brcmstb: use BRCMSTB-specific options only if supported")
Signed-off-by: Florian Fainelli <fainelli@broadcom.com>