busybox: devise a way for applying generic *and* version patches
Buildroot supports two mechanisms for applying patches to a package:
  1. generic patches
  2. version specific patches

Generic patches are meant for any version of said package. The same
patches will continue to be applied if the package version changes.

Version specific patches, which live in a sub-directory named after the
package version, are applied only to that specific version of
the package.

The problem is that these two methods are mutually exclusive. A package
either receives all generic patches (if there is no version directory)
or it will *only* receive the version specific patches and none of the
generic ones.

For busybox, we want to be able to apply custom, version specific
patches while retaining all the public, generic patches, i.e. we need
to combine both methods. Therefore, we introduce a pre-patch hook that
applies the version specific patches first and relies on the standard
Buildroot patching mechanism to apply the generic ones.

Our version specific patches live in a separate sub-directory named
brcmstb-$(VERSION) to prevent the standard mechanism for version
specific patches (as detailed above) from kicking in.

refs #SWLINUX-5670

Signed-off-by: Markus Mayer <mmayer@broadcom.com>