br_config.pl: warn if kernel remote is not accessible
Print a warning if we can't resolve the name for the server hosting
our Linux repo. Also, don't try to access said remote repo to determine
the kernel's SHA, since that wouldn't work anyway.

The warning is not displayed if a local Linux directory has been
specified and we don't need the remote Linux repo for anything.

refs #SWLINUX-5696

Signed-off-by: Markus Mayer <mmayer@broadcom.com>