br_config.pl: sanity-check the IP of the Buildroot mirror
We need to be more diligent about verifing the IP address of the
specified Buildroot mirror.

It is possible that a DNS service provider will return a valid IP
address for any and and all hostnames, even those it can't actually
resolve. This may be intended as catch-all to redirect the user to a
search engine or some similar service.

However, in our case, such a "false postive" response will lead us to
think we can use the internal Buildroot mirror to download sources
when, in reality, the hostname will resolve to a completely different
machine that isn't a Buildroot mirror at all.

So, let's make sure the IP address for the Buildroot mirror is on one
of these "private" networks before we try to use it:
	10.0.0.0    - 10.255.255.255
	172.14.0.0  - 172.31.255.255
	192.168.0.0 - 192.168.255.255

refs #SWLINUX-5717

Signed-off-by: Markus Mayer <mmayer@broadcom.com>