br_config.pl: introduce new option "-C" to query compiler version
With the removal of misc/toolchain, third party scripts may need a
different way of determining which toolchain to use.

We provide this functionality via the new command line option "-C".
After determining the recommended toolchain, br_config.pl will then
print the version and exit.

Therefore old code like
	buildroot_toolchain=`cat misc/toolchain`
can be replaced with a variant of
	buildroot_toolchain=`bin/br_config.pl -C -L ../linux arm`

"br_config.pl -C" does require the architecture to be specified in case
we'll ever want to use a different toolchain based on the target
architecture.

Other arguments, such as -L or -v, can be specified with -C and will
influence the toolchain selection.

refs #SWLINUX-5741

Signed-off-by: Markus Mayer <mmayer@broadcom.com>