br_config.pl: add support for LLVM toolchains
We add support for LLVM as external toolchain. At this time, the LLVM
toolchain needs to be accompanied by a GCC toolchain for the sys-root
as well as some low-level linker files.

We also add a kernel version check. LLVM is only supported as of STB
Linux 5.4. Earlier kernels (such as STB Linux 4.9) will likely result
in build errors.

refs #SWLINUX-5553

Signed-off-by: Markus Mayer <mmayer@broadcom.com>