package/cpufrequtils: Get it to build with Clang/LLVM
utils.h is unconditionally included even if most of the routines are not
used and are properly gated with the __i386__ and __x86_64__
conditionals.

The "a" constraint is not valid while building for ARM or ARM64, just
provide an empty stub when not building for i386 or x86-64 since that
code is just not functional anyway.

refs #SWLINUX-5553

Signed-off-by: Florian Fainelli <florian.fainelli@broadcom.com>