package/xfsprogs: provide a patch to limit ".o" -> ".lo" substitutions
To prevent issues if the ".o" extension appears in a directory path,
ensure that the ".o" -> ".lo" substitution is only performed for the
final file extension of a path.

As discussed in [1], we can't match against "\.o$" but must use "\.o:".
That's because a string as shown in the example below needs the proper
occurrence of ".o" (i.e. the one preceeding the colon) replaced with
".lo".

Before substitution:
    foo.o/bitmap.o: bitmap.c ../include/xfs.h ../include/xfs/linux.h

After substitution:
    foo.o/bitmap.lo: bitmap.c ../include/xfs.h ../include/xfs/linux.h

Using "\.o$" in the regex wouldn't replace anything.

[1] https://www.spinics.net/lists/linux-xfs/msg49712.html

refs #SWLINUX-5803

Signed-off-by: Markus Mayer <mmayer@broadcom.com>