br_config.pl: introduce ability to check for AMS tracing remotely
Introduce the ability to query a remote GIT repo for the presence of
files that indicate support for the AMS tracing feature.

This ability allows us to accurately detect whether or not the selected
kernel supports the AMS tracing feature, even if the kernel sources
sill need to be cloned from a remote repo.

The approach of checking a remote GIT repository for the existence of a
file requires support from the GIT server. Specifically, the config
option "daemon.uploadarch" must be set to "true" on the server-side
repo.

refs #SWLINUX-5789

Signed-off-by: Markus Mayer <mmayer@broadcom.com>