package/libunwind: Fixed Clang builds
The AArch64 build would be failing because libunwind picked up the
liblzma dependency and it would not be automatically resolved while the
configure tests for libunwind (this smells like a pkgconfig issue).
Disabling "minidebuginfo" removes the dependency on liblzma.

The ARM build would be succeeding but later on the strace configure
script would fail due to a lack of a pid_t definition.

refs #SWLINUX-5553

Signed-off-by: Florian Fainelli <florian.fainelli@broadcom.com>