board: brcmstb: post-image: copy vmlinux & co to aid with debugging
It can be helpful to have vmlinux and, if they exist, additional symbol
or annotation files, available to aid with debugging.

By copying these files into the images/debug directory, they will be
available in the Jenkins build artifacts. They'll also be easier to
access in a well-known location after a manual build.

Beware that these files do NOT contain the target platform in their
file name (unlike the vmlinuz images and the NFS tar-ball). The
platform can be derived by the directory path, however
(output/$platform/images/debug/).

refs #SWLINUX-5942

Signed-off-by: Markus Mayer <mmayer@broadcom.com>