br_config.pl: determine the full version of the kernel being used
Detect the full version of the kernel that is being used, irrespective
of whether it is a remote source tree or a local one. Display this
information for the user.

Bail if we can't determine the version number, since that means
something undesirable is going on with the top-level Linux makefile.
Building likely wouldn't work anyway.

refs #SWLINUX-6002

Signed-off-by: Markus Mayer <mmayer@broadcom.com>