br_config.pl: disable ubihealthd (part of mtd-utils) for Linux < 5.1
As per the initial ubihealthd commit message, ubihealthd requires at
least kernel 5.1 to be usable:

http://git.infradead.org/mtd-utils.git/commit/7f0e2dc21fb2

In addition, we started to run into build issues with ubihealthd as of
Buildroot 2021.08 when trying to build with stbgcc-6.3. We are using
stbgcc-6.3 for our 4.x kernels only. So, let's disable ubihealthd for
those kernels to avoid the issue, since ubihealthd wouldn't work there
anyway.

The build issue itself was caused by the absence of getrandom() in
glibc-2.24, which is bundled with stbgcc-6.3. The problem does not
exist with stbgcc-8.3, since that comes with glibc-2.27, and glibc-2.27
does provide the getrandom() function.

refs #SWLINUX-6002

Signed-off-by: Markus Mayer <mmayer@broadcom.com>